requirements.txt
# configure async fastapi
fastapi
gunicorn
uvicorn
# use postres database backend
fastapi-sqlalchemy
pydantic
alembic
psycopg2
python-dotenv
# config async
databases
databases[postgresql]
asyncpg
